@@ -5,24 +5,30 @@
   "main": "index.js",
   "dependencies": {},
   "devDependencies": {
-    "@comparaonline/eslint-config-react": "^1.0.3",
     "babel-core": "^6.21.0",
     "babel-jest": "^19.0.0",
     "babel-loader": "^6.2.10",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.24.0",
-    "babel-preset-es2015": "^6.24.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.23.0",
-    "eslint": "^3.12.2",
-    "jest": "^19.0.2",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-15": "^1.0.5",
+    "eslint": "^4.18.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.9.0",
+    "eslint-plugin-jest": "^21.12.2",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.7.0",
+    "jest": "^22.4.0",
     "prop-types": "^15.5.10",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2",
+    "react-test-renderer": "^15.6.2",
     "webpack": "^2.2.1",
     "webpack-cleanup-plugin": "^0.5.1"
   },
   "peerDependencies": {
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2"
+    "react": "^15.6.2",
+    "react-dom": "^15.6.2"
   },
   "scripts": {
     "lint": "eslint ./src",
@@ -46,5 +52,10 @@
   "bugs": {
     "url": "https://github.com/comparaonline/chat-component/issues"
   },
-  "homepage": "https://github.com/comparaonline/chat-component#readme"
+  "homepage": "https://github.com/comparaonline/chat-component#readme",
+  "jest": {
+    "setupFiles": [
+      "./src/config/jest.config.js"
+    ]
+  }
 }
